refactor(npu): simplify async CAM ubatching
